Mover algumas funções para common.mjs
